Refactor: clean the page
Extract components, hooks and functions to separate files
